requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
asyncpg==0.29.0
alembic==1.13.1
pydantic-settings==2.1.0
python-multipart==0.0.9
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
uuid6==2024.1.12
